Fixed for 64bit archs

+++ main/SAPI.c
@@ -18,7 +18,7 @@
    +----------------------------------------------------------------------+
 */
 
-/* $Id: SAPI.c,v 1.178 2003/09/10 08:47:37 sr Exp $ */
+/* $Id: SAPI.c,v 1.179 2003/09/13 15:27:04 abies Exp $ */
 
 #include <ctype.h>
 #include <sys/stat.h>
@@ -517,7 +517,7 @@ SAPI_API int sapi_header_op(sapi_header_op_enum op, void *arg TSRMLS_DC)
 
 	switch (op) {
 	case SAPI_HEADER_SET_STATUS:
-		sapi_update_response_code((int) arg TSRMLS_CC);
+		sapi_update_response_code((long) arg TSRMLS_CC);
 		return SUCCESS;
 
 	case SAPI_HEADER_REPLACE:
